add Personal Account model
This commit is a little bit bigger than others that only add model because we also created the Factory. Which is used by the test suite to easily test Personal Account.

It could be two separate commits to keep the diff small and easier to review, but the intent is to have only green commits (commits that pass the whole test suite) in the master branch.