change balance to a derived column in personal_accounts
After trying to do the transfer operation, it was evident that a approach that favors consistency would be more desired. This is the start of that approach.

Instead of having the transfer done inside a database transaction, and having to reason about if the isolation level should be Y or Z, we can change the balance to a derived column updated by a trigger, and attach a constraint to that column so that the balance never goes lower than zero.

The trade-off here is that the logic about balance is inside a trigger, but I believe it's the right trade-off as we're instead having a super strong guarantee of data integrity